refactor(login): Separate into components
refs: #60
