+++ webapp/client/src/domain/api/routes/components/stats.js
