+++ webapp/client/src/presenter/features/auth/login/loginButton.jsx
@@ -1,0 +1,14 @@
+function LoginButton({ isLoading }) {
+    return (
+      <button
+        type="submit"
+        id="loginButton"
+        className={isLoading ? "load" : ""}
+      >
+        {isLoading ? "Completed" : "Verify"}
+      </button>
+    );
+  }
+  
+  export default LoginButton;
+
