+++ webapp/client/src/presenter/features/auth/login/logoSection.jsx
@@ -1,0 +1,11 @@
+function LogoSection() {
+    return (
+      <section id="logoLoginPage">
+        <img src="/images/droneBottom.svg" alt="Drone Tech Logo" />
+        <h1>Drone Tech</h1>
+      </section>
+    );
+  }
+  
+  export default LogoSection;
+
