+++ webapp/client/src/presenter/features/auth/login/registerLink.jsx
@@ -1,0 +1,14 @@
+function RegisterLink({ onClick }) {
+    return (
+      <p>
+        Don't have an account?
+        <span id="registerUser" onClick={onClick}>
+          {" "}
+          Register
+        </span>
+      </p>
+    );
+  }
+  
+  export default RegisterLink;
+
