+++ webapp/client/src/presenter/features/core/burgerMenu.jsx
@@ -1,4 +1,3 @@
-import '../styles.css';
 import { Link } from 'react-router-dom'
 
 function BurgerMenu({isVisible})
