+++ webapp/client/src/presenter/features/core/infoBlock.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import NavMenu from './NavMenu';
-import "../styles.css";
+import NavMenu from '../../layout/navMenu';
 
 function MainContent() {
   return (
