+++ webapp/client/src/presenter/features/core/loadingScreen.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect } from "react";
-import "../styles.css";
 
 const LoadingScreen = () => {
   useEffect(() => {
